[tests] Add SAMPLE_APP
