+++ openwisp_monitoring/device/tests/test_settings.py
@@ -1,31 +1,10 @@
-from unittest.mock import patch
+import os
+from unittest import skipIf
 
-from django.core.exceptions import ImproperlyConfigured
-
+from ..base.tests.test_settings import BaseTestSettings
 from . import DeviceMonitoringTestCase
 
 
-class TestSettings(DeviceMonitoringTestCase):
-    """
-    Tests ``OpenWISP Device settings`` functionality
-    """
-
-    @patch(
-        'django.conf.settings.OPENWISP_MONITORING_CRITICAL_DEVICE_METRICS',
-        [{}],
-        create=True,
-    )
-    def test_invalid_critical_device_metrics_setting(self):
-        with self.assertRaises(ImproperlyConfigured):
-            from ..settings import get_critical_device_metrics
-
-            get_critical_device_metrics()
-
-    @patch(
-        'django.conf.settings.OPENWISP_MONITORING_HEALTH_STATUS_LABELS', {}, create=True
-    )
-    def test_invalid_health_status_setting(self):
-        with self.assertRaises(ImproperlyConfigured):
-            from ..settings import get_health_status_labels
-
-            get_health_status_labels()
+@skipIf(os.environ.get('SAMPLE_APP', False), 'Running tests on SAMPLE_APP')
+class TestSettings(BaseTestSettings, DeviceMonitoringTestCase):
+    app_name = 'device_monitoring'
