+++ tests/openwisp2/sample_check/apps.py
@@ -1,0 +1,6 @@
+from openwisp_monitoring.check.apps import CheckConfig
+
+
+class SampleCheckConfig(CheckConfig):
+    name = 'openwisp2.sample_check'
+    label = 'sample_check'
