+++ tests/openwisp2/sample_check/models.py
@@ -1,0 +1,8 @@
+from openwisp_monitoring.check.base.models import AbstractCheck
+from swapper import swappable_setting
+
+
+class Check(AbstractCheck):
+    class Meta(AbstractCheck.Meta):
+        abstract = False
+        swappable = swappable_setting('check', 'Check')
