+++ tests/openwisp2/sample_device_monitoring/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'openwisp_monitoring.device.apps.DeviceMonitoringConfig'
